se cambio en la entity y el request de PiezaDental el numero de pieza de bigdecimal a integer porque era inesesario

+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Model/PiezaDental.java
@@ -20,7 +20,7 @@ public class PiezaDental {
     private Long idPiezaDental; // No auto-generado: número de pieza es fijo (1-32)
 
     @Column(name = "numero_pieza", nullable = false, precision = 5, scale = 2)
-    private BigDecimal numeroPieza;
+    private Integer numeroPieza;
 
     @Column(name = "estado", length = 50)
     private String estado;
